+++ src/library/DIPS.Mobile.UI/Components/Images/Image/iOS/ImageHandler.cs
@@ -1,3 +1,4 @@
+using DIPS.Mobile.UI.Internal.Logging;
 using Microsoft.Maui.Platform;
 using UIKit;
 
@@ -14,15 +15,23 @@ public partial class ImageHandler
             // Wait for Image to be set
             while (handler.PlatformView.Image is null)
             {
-                if(tries > 10)
+                if (tries > 10)
                     return;
-                
+
                 await Task.Delay(1);
                 tries++;
             }
-            
+
             if (image.TintColor == null) return;
-            handler.PlatformView.Image = handler.PlatformView.Image.ImageWithRenderingMode(UIImageRenderingMode.AlwaysTemplate);
+
+            if (handler.PlatformView is null)
+            {
+                DUILogService.LogError<ImageHandler>("PlatformView is null, this should not happen, most likely the issue is that the Content is rendered and then the handler is instantly disconnected");
+                return;
+            }
+
+            handler.PlatformView.Image =
+                handler.PlatformView.Image.ImageWithRenderingMode(UIImageRenderingMode.AlwaysTemplate);
             handler.PlatformView.TintColor = image.TintColor.ToPlatform();
         }
         catch